add button to scroll messages into view
