+++ src/components/Chat.js
@@ -2,21 +2,68 @@ import TopNavigationBar from "./TopNavigationBar"
 import LeftSideComponent from "./LeftSideComponent"
 import MiddleChatWindow from "./MiddleChatWindow"
 import BottomInputComponent from "./BottomInputComponent"
-import { useEffect } from "react"
+import { db } from "../config/fire-config"
+import { collection, onSnapshot, query, orderBy } from "firebase/firestore"
+import { useEffect, useContext, useState, useRef, useMemo } from "react"
 import { useNavigate } from "react-router-dom"
 import { ThemeContext } from "../context.js"
-import { useContext } from "react"
 import { SoundProvider } from "../context/SoundContext"
 
 function Chat({ currentUser }) {
+  const [messages, setMessages] = useState([])
+  const messagesCollectionRef = collection(db, "messages")
+  const screenBottom = useRef(null)
+  const lastMessageIsInViewport = useIsInViewport(screenBottom)
+  const queryMessages = query(
+    messagesCollectionRef,
+    orderBy("timestamp", "desc")
+  )
   const navigate = useNavigate()
   const { theme } = useContext(ThemeContext)
+
+  const getMessages = () => {
+    onSnapshot(queryMessages, (snapshot) => {
+      setMessages(snapshot.docs.map((doc) => ({ ...doc.data(), id: doc.id })))
+    })
+  }
+
+  const scrollToBottom = () => {
+    screenBottom.current?.scrollIntoView({ behavior: "smooth" })
+  }
+
+  function useIsInViewport(ref) {
+    const [isIntersecting, setIsIntersecting] = useState(false)
+
+    const observer = useMemo(
+      () =>
+        new IntersectionObserver(([entry]) => {
+          setIsIntersecting(entry.isIntersecting)
+        }),
+      []
+    )
+
+    useEffect(() => {
+      observer.observe(ref.current)
+      return () => {
+        observer.disconnect()
+      }
+    }, [ref, observer])
+    return isIntersecting
+  }
 
   useEffect(() => {
     if (!currentUser) {
       navigate("/")
+    } else {
+      getMessages()
     }
   }, [])
+
+  useEffect(() => {
+    if (lastMessageIsInViewport) {
+      scrollToBottom()
+    }
+  }, [messages])
 
   return (
     <>
@@ -31,13 +78,22 @@ function Chat({ currentUser }) {
               <LeftSideComponent />
             </div>
             <div className='col-8'>
-              <MiddleChatWindow currentUser={currentUser} />
+              <MiddleChatWindow
+                currentUser={currentUser}
+                messages={messages}
+                screenBottom={screenBottom}
+              />
             </div>
             <div className='col-2'></div>
           </div>
           {/* Provide the global sound context value to this component */}
           <SoundProvider>
-            <BottomInputComponent currentUser={currentUser} isFocused />
+            <BottomInputComponent
+              currentUser={currentUser}
+              lastMessageIsInViewport={lastMessageIsInViewport}
+              scrollToBottom={scrollToBottom}
+              isFocused
+            />
           </SoundProvider>
         </div>
       </section>
